updates dynamic import setup
