+++ utils/utils/package.json
@@ -4,6 +4,9 @@
   "main": "dist/index.js",
   "author": "Tyler Benton",
   "license": "MIT",
+  "scripts": {
+    "postinstall": "node ./post-install.js"
+  },
   "publishConfig": {
     "access": "public"
   },
