--- utils/utils/packages.json
@@ -1,10 +1,0 @@
-[
-  "apollo",
-  "credit-cards",
-  "files",
-  "haptic",
-  "is",
-  "locale",
-  "react",
-  "react-native"
-]
