+++ utils/utils/src/packages.mjs
@@ -1,49 +1,2 @@
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-// !!! This file is generated do not modify it manually because it will be overwritten
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-// is this stupid to do. Yes. But unfortuantly you have to require packages
-// specifically, you can't use any variables in a require function in react-native
-/* eslint-disable global-require, no-empty, import/no-extraneous-dependencies */
-
-try {
-  const pkg = require('@ma-shop/apollo')
-  if (pkg) {
-    Object.assign(module.exports, pkg)
-  }
-} catch (e) { }
-
-try {
-  const pkg = require('@ma-shop/haptic')
-  if (pkg) {
-    Object.assign(module.exports, pkg)
-  }
-} catch (e) { }
-
-try {
-  const pkg = require('@ma-shop/is')
-  if (pkg) {
-    Object.assign(module.exports, pkg)
-  }
-} catch (e) { }
-
-try {
-  const pkg = require('@ma-shop/locale')
-  if (pkg) {
-    Object.assign(module.exports, pkg)
-  }
-} catch (e) { }
-
-try {
-  const pkg = require('@ma-shop/react')
-  if (pkg) {
-    Object.assign(module.exports, pkg)
-  }
-} catch (e) { }
-
-try {
-  const pkg = require('@ma-shop/react-native')
-  if (pkg) {
-    Object.assign(module.exports, pkg)
-  }
-} catch (e) { }
+// this is done to prevent unnecessary changes to the `@ma-shop/utils` package
+export * from '../packages'
